+++ Backend/src/main/java/com/softwareProject/banksApplication/dto/response/invoice/InvoiceResponse.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class InvoiceResponse {
     private Long id;
-    private int invoiceNo;
+    private Long invoiceNo;
     private InvoiceInfo.InvoiceType invoiceType;
     private BigDecimal invoiceAmount;
     private LocalDateTime payDate;
